+++ client/src/features/account/pages/AccountSettingsPage.jsx
@@ -1,0 +1,15 @@
+import { AccountSettingsForm } from "../components/AccountSettingsForm";
+
+export function AccountSettingsPage() {
+  return (
+    <div className="flex flex-1 flex-col">
+      <div className="@container/main flex flex-1 flex-col gap-2">
+        <div className="flex flex-col gap-4 py-4 md:gap-6 md:py-6">
+          <div className="px-4 lg:px-6">
+            <AccountSettingsForm />
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}